Clean various code
* Check pep8
* Sort import
* Remove empty/unused file
* Remove unused views/functions
