+++ web/Web/web/core/signals.py
@@ -1,5 +1,5 @@
-from django.db.models.signals import post_delete, post_save
-from django.db.models.signals import post_init
+from django.db.models.signals import post_delete
+from django.db.models.signals import post_save
 from django.dispatch import receiver
 from notifications.signals import notify
