+++ web/Web/web/interfaces/SearchEngine/anserini.py
@@ -1,8 +1,9 @@
-import requests
-
 from config.settings.base import DEFAULT_NUM_DISPLAY
 from config.settings.base import SEARCH_SERVER_IP
 from config.settings.base import SEARCH_SERVER_PORT
+
+import requests
+
 from web.interfaces.SearchEngine.base import SearchInterface
