+++ web/Web/web/interfaces/SearchEngine/base.py
@@ -1,4 +1,5 @@
-from abc import ABC, abstractmethod
+from abc import ABC
+from abc import abstractmethod
 
 
 class SearchInterface(ABC):
