+++ web/Web/web/interfaces/SearchEngine/elastic.py
@@ -1,11 +1,11 @@
-import json
-
-import requests
-
 from config.settings.base import DEFAULT_NUM_DISPLAY
 from config.settings.base import INDEX_NAME
 from config.settings.base import SEARCH_SERVER_IP
 from config.settings.base import SEARCH_SERVER_PORT
+import json
+
+import requests
+
 from web.interfaces.SearchEngine.base import SearchInterface
