+++ web/Web/web/judgment/urls.py
@@ -1,4 +1,5 @@
-from django.urls import path, re_path
+from django.urls import path
+from django.urls import re_path
 
 from web.judgment import views
 
@@ -10,13 +11,13 @@ urlpatterns = [
          name='view'),
 
     # Ajax views
-    path(r'post_judgment/', views.JudgmentAJAXView.as_view(),
+    path('post_judgment/', views.JudgmentAJAXView.as_view(),
          name='post_judgment'),
-    path(r'post_nojudgment/', views.NoJudgmentAJAXView.as_view(),
+    path('post_nojudgment/', views.NoJudgmentAJAXView.as_view(),
          name='post_nojudgment'),
     re_path(r'^get_latest/(?P<number_of_docs_to_show>\d+)/$',
             views.GetLatestAJAXView.as_view(),
             name='get_latest'),
-    path(r'^get_all/', views.GetAllAJAXView.as_view(),
-          name='get_all'),
+    path('get_all/', views.GetAllAJAXView.as_view(),
+         name='get_all'),
 ]
