--- web/Web/web/time/Idle.js/core.html
@@ -1,29 +1,0 @@
-
-<html>
-	<head>
-		<title>Idle.Js Test</title>
-		<script src="http://ajax.googleapis.com/ajax/libs/jquery/1.7.1/jquery.min.js"></script>
-		<script src="./idle.js"></script>
-		<script src="./timer.js"></script>
-		<script>
-        var foo = AwayTimer();
-
-		</script>
-		<style>
-			#ActivityList {
-				border: 1px solid #ccc;
-				padding: 5px;
-				width: 400px;
-			}
-			#ActivityList li {
-				border-bottom: 1px solid #eee;
-				margin: 5px 0;
-			}
-		</style>
-	</head>
-	<body>
-		<!-- <label for="Timeout">Set timeout</label>
-		<input id="Timeout" type="text" value="5000" /> -->
-		<ul id="ActivityList"></ul>
-	</body>
-</html>
